Use NSUInteger instead of uint_fast32_t, as the generated assembly is identical and NSUInteger is a more familiar type to ObjC developers.

+++ LiveFrost/LFGlassView.h
@@ -8,6 +8,6 @@
 
 @property (nonatomic, assign, getter=isLiveBlurring) BOOL liveBlurring;
 
-@property (nonatomic, assign) uint_fast32_t frameInterval;
+@property (nonatomic, assign) NSUInteger frameInterval;
 
 @end
